requirements.txt
azure_storage==0.37.0
Pillow==10.4.0
PyYAML==6.0.2
streamlit==1.38.0
sympy==1.13.1
